+++ hypervisor/Makefile
@@ -154,6 +154,7 @@ C_SRCS += arch/x86/virq.c
 C_SRCS += arch/x86/vmexit.c
 C_SRCS += arch/x86/vmx.c
 C_SRCS += arch/x86/vmcs.c
+C_SRCS += arch/x86/virtual_cr.c
 C_SRCS += arch/x86/assign.c
 C_SRCS += arch/x86/trusty.c
 C_SRCS += arch/x86/cpu_state_tbl.c
