+++ hypervisor/include/arch/x86/guest/vcpu.h
@@ -53,6 +53,7 @@
 #ifndef ASSEMBLER
 
 #include <guest.h>
+#include <virtual_cr.h>
 
 /**
  * @brief vcpu
@@ -416,69 +417,6 @@ uint64_t vcpu_get_rflags(struct acrn_vcpu *vcpu);
  * @param[in] val the value set RFLAGS
  */
 void vcpu_set_rflags(struct acrn_vcpu *vcpu, uint64_t val);
-
-/**
- * @brief get vcpu CR0 value
- *
- * Get & cache target vCPU's CR0 in run_context.
- *
- * @param[in] vcpu pointer to vcpu data structure
- *
- * @return the value of CR0.
- */
-uint64_t vcpu_get_cr0(struct acrn_vcpu *vcpu);
-
-/**
- * @brief set vcpu CR0 value
- *
- * Update target vCPU's CR0 in run_context.
- *
- * @param[inout] vcpu pointer to vcpu data structure
- * @param[in] val the value set CR0
- */
-void vcpu_set_cr0(struct acrn_vcpu *vcpu, uint64_t val);
-
-/**
- * @brief get vcpu CR2 value
- *
- * Get & cache target vCPU's CR2 in run_context.
- *
- * @param[in] vcpu pointer to vcpu data structure
- *
- * @return the value of CR2.
- */
-uint64_t vcpu_get_cr2(struct acrn_vcpu *vcpu);
-
-/**
- * @brief set vcpu CR2 value
- *
- * Update target vCPU's CR2 in run_context.
- *
- * @param[inout] vcpu pointer to vcpu data structure
- * @param[in] val the value set CR2
- */
-void vcpu_set_cr2(struct acrn_vcpu *vcpu, uint64_t val);
-
-/**
- * @brief get vcpu CR4 value
- *
- * Get & cache target vCPU's CR4 in run_context.
- *
- * @param[in] vcpu pointer to vcpu data structure
- *
- * @return the value of CR4.
- */
-uint64_t vcpu_get_cr4(struct acrn_vcpu *vcpu);
-
-/**
- * @brief set vcpu CR4 value
- *
- * Update target vCPU's CR4 in run_context.
- *
- * @param[inout] vcpu pointer to vcpu data structure
- * @param[in] val the value set CR4
- */
-void vcpu_set_cr4(struct acrn_vcpu *vcpu, uint64_t val);
 
 /**
  * @brief get guest emulated MSR
